Prevent early publication of unofficial results
This was discussed in an inference meeting, and appeared to be broadly considered legal. However, it doesn't seem healthy for MLPerf: it allows for gaming of publication (e.g. a submitter, knowing the official numbers, could publish day-zero updated unofficial numbers), and erases some of the the advantage that official submitters should reasonably have over unofficial ones.