Add DC inference sys JSON networking requirements
Add host/accelerator networking requirements for datacenter inference submission systems.  
System networking description is required for inference datacenter submissions because a datacenter 
system submitting to MLPerf inference must have at least enough networking bandwidth either on the
host or on the device to sustain the server scenario throughput, and it is important to populate the networking
especially during power measurements.  Refer to anirban-ghosh-patch-2 in inference_policies for more details.