Update to mandatory fields : System Json description
Power submissions require power to be measured at the system level .  The current System Json description does not have sufficient required response fields to capture the full system description for giving visibility into the nature of the system submitted. After reviews in the Power WG , we are submitting this update to the "meaningful response required" fields.   We'd like to use this for the MLPerf Inference v2.0 submissions.